refactor: remove static module page and update imports
